stablediffusion/requirements.txt
# Core dependencies for Watermarked Stable Diffusion WebUI

# AI/ML frameworks
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.20.0
transformers>=4.25.0

# Web interface
flask>=2.0.0

# Image processing
pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.7.0

# Cryptography for watermarking
cryptography>=3.4.0

# Scientific computing
scipy>=1.9.0

# Utilities
omegaconf>=2.3.0
